use uint64_t for result counts instead of uint32_t in C API

+++ include/capi/IdVisitor.h
@@ -31,14 +31,14 @@ class IdVisitor : public SpatialIndex::IVisitor
 {
 private:
     std::vector<uint64_t> m_vector;
-    uint32_t nResults;
+    uint64_t nResults;
 
 public:
 
     IdVisitor();
     ~IdVisitor();
 
-    uint32_t GetResultCount() const { return nResults; }
+    uint64_t GetResultCount() const { return nResults; }
     std::vector<uint64_t>& GetResults()  { return m_vector; }
     
     void visitNode(const SpatialIndex::INode& n);
